Adjusted theme. Brightened up site title.
Brightened up text and title

Changed some colors around.

Adjusted posts

Changed the highlighting in the post to something a bit easier to see
and copy/paste.

Added avatar and homepage description

Added a picture of my mug to the home page and I also put a welcome
message too for people to read upon arrival. I figured out the syntax
for highlighting the powershell objects too and changed that in the caps
to control post.

Set up post previews

Set up the blog so that on the homepage when hovering over a post's link
it will give you a 10 word preview of the post. I also set the archive
page up to have a preview under each archived post along with a link to
read more.